Add inventory2

+++ DirectX/Engine/zzAnimation.cpp
@@ -27,7 +27,7 @@ namespace zz
         if (mbComplete)
             return;
 
-        mTime += Time::DeltaTime();
+        mTime += (float)Time::DeltaTime();
 
         if (mSprites[mIndex].duration <= mTime)
         {
@@ -36,7 +36,7 @@ namespace zz
 
             if (mSprites.size() <= mIndex)
             {
-                mIndex = mSprites.size() - 1;
+                mIndex = (int)(mSprites.size() - 1);
                 mbComplete = true;
             }
         }
